+++ src/hopsy/misc.py
@@ -1,6 +1,7 @@
 """
 
 """
+import hopsy
 
 
 class _core:
@@ -68,6 +69,8 @@ def MarkovChain(
         elif problem.starting_point is not None:
             _proposal = proposal(problem, starting_point=problem.starting_point)
         else:
+            # _starting_point = hopsy.compute_chebyshev_center(problem)
+            # _proposal = proposal(problem, _starting_point)
             _proposal = proposal(problem)
     else:
         _proposal = proposal