+++ src/hopsy/proposal.hpp
@@ -539,6 +539,8 @@ namespace hopsy {
             if (problem) {
                 if (!problem->startingPoint) {
                     throw std::runtime_error("Cannot setup a proposal without starting point.");
+//                    VectorType starting_point = computeChebyshevCenter(*problem);
+//                    return UninitializedProposalWrapper<ProposalImpl, Args...>(problem->A, problem->b, startingPoint, args...);
                 }
                 return UninitializedProposalWrapper<ProposalImpl, Args...>(problem->A, problem->b, *problem->startingPoint, args...);
             } else {
@@ -552,6 +554,8 @@ namespace hopsy {
             if (problem) {
                 if (!(problem->startingPoint || startingPoint)) {
                     throw std::runtime_error("Cannot setup a proposal without starting point.");
+//                    VectorType starting_point = computeChebyshevCenter(*problem);
+//                    return UninitializedProposalWrapper<ProposalImpl, Args...>(problem->A, problem->b, startingPoint, args...);
                 }
                 VectorType _startingPoint = ( startingPoint ? *startingPoint : *problem->startingPoint );
                 return UninitializedProposalWrapper<ProposalImpl, Args...>(problem->A, problem->b, _startingPoint, args...);